+++ lib/features/record/components/announcement_bar/announcement_bar.dart
@@ -16,7 +16,7 @@ import 'package:pilll/features/record/components/announcement_bar/components/res
 import 'package:pilll/features/sign_in/sign_in_sheet.dart';
 import 'package:pilll/provider/locale.dart';
 import 'package:pilll/provider/premium_and_trial.codegen.dart';
-import 'package:pilll/provider/typed_shared_preferences.dart';
+import 'package:pilll/provider/old_typed_shared_preferences.dart';
 import 'package:pilll/provider/auth.dart';
 import 'package:pilll/utils/datetime/day.dart';
 import 'package:pilll/utils/shared_preference/keys.dart';
